LayoutTests/imported/w3c/web-platform-tests/css/selectors/invalidation/input-pseudo-classes-in-has-expected.txt

 Check me!

PASS :checked & :indeterminate invalidation on <input>
PASS :indeterminate invalidation on <progress>
PASS :disabled invalidation
PASS :read-only invalidation
PASS :valid invalidation
FAIL :default invalidation with input[type=radio] assert_equals: ancestor should be lightblue expected "rgb(173, 216, 230)" but got "rgb(0, 0, 0)"
PASS :required invalidation
FAIL :out-of-range invalidation assert_equals: ancestor should be darkgreen expected "rgb(0, 100, 0)" but got "rgb(0, 0, 0)"
PASS :placeholder-shown invalidation
